python/PLZA-ShinyHunter/src/requirements.txt
opencv-python
keyboard
pyserial